Refactor: Update schedule lessons list with DiffUtil
This commit refactors the way the list of lessons in `ScheduleFragment` is updated. Instead of clearing and repopulating the entire list and notifying the adapter, it now uses `DiffUtil` to calculate the differences between the old and new lists. This allows for more efficient updates to the `RecyclerView`, only updating the items that have changed, instead of reloading all the items.

-   Created `LessonsDiffCallback` class for calculating the difference between two lists of `ScheduleListLesson`.
-   Implemented `DiffUtil.calculateDiff` in `ScheduleFragment` to compute the difference between old and new list.
-   Updated the adapter with the diff results using `dispatchUpdatesTo` in `ScheduleFragment`.
- Removed the whole clearing and adding in `lessonsMutableLiveData` observer.